# docs/swagger.yaml
definitions:
  _const.MailType:
    enum:
    - plain mail
    - registered mail
    - plain postcard
    - registered postcard
    - plain presswork
    - registered presswork
    - post package
    - post express package
    - post small package
    type: string
    x-enum-comments:
      MailTypeExpressPackage: 包裹：快递
      MailTypePackage: 包裹：一般
      MailTypePlainMail: 信件：一般
      MailTypePlainPostcard: 明信片：一般
      MailTypePlainPresswork: 印刷品：一般
      MailTypeRegisteredMail: 信件：挂号
      MailTypeRegisteredPostcard: 明信片：挂号
      MailTypeRegisteredPresswork: 印刷品：挂号
      MailTypeSmallPackage: 包裹：小包
    x-enum-varnames:
    - MailTypePlainMail
    - MailTypeRegisteredMail
    - MailTypePlainPostcard
    - MailTypeRegisteredPostcard
    - MailTypePlainPresswork
    - MailTypeRegisteredPresswork
    - MailTypePackage
    - MailTypeExpressPackage
    - MailTypeSmallPackage
  _const.PostFeePaymentType:
    enum:
    - cash
    - stamps
    type: string
    x-enum-comments:
      PostFeePaymentTypeCash: 现金
      PostFeePaymentTypeStamps: 邮票
    x-enum-varnames:
    - PostFeePaymentTypeCash
    - PostFeePaymentTypeStamps
  _const.PostRegion:
    enum:
    - domestic local
    - domestic
    - overseas (by air)
    - overseas (surface)
    - overseas (by air surface)
    type: string
    x-enum-comments:
      PostRegionDomestic: 国内外埠
      PostRegionDomesticLocal: 国内本埠
      PostRegionOverseasByAir: 跨洋航空
      PostRegionOverseasByAirAndSurface: 跨洋航空水陆路
      PostRegionOverseasSurface: 跨洋水陆路
    x-enum-varnames:
    - PostRegionDomesticLocal
    - PostRegionDomestic
    - PostRegionOverseasByAir
    - PostRegionOverseasSurface
    - PostRegionOverseasByAirAndSurface
  _const.RecipientType:
    enum:
    - myself
    - couple
    - relative
    - friend
    - pen pal
    - postmark collecting
    - other
    type: string
    x-enum-varnames:
    - RecipientTypeMyself
    - RecipientTypeCouple
    - RecipientTypeRelative
    - RecipientTypeFriend
    - RecipientTypePenPal
    - RecipientTypePostmarkCollecting
    - RecipientTypeOther
  form.AddContactForm:
    properties:
      address:
        example: CaotangLu 1 Hao
        type: string
      district:
        example: SichuanSheng ChengduShi QingyangQu
        type: string
      name:
        example: Leon Kennedy
        type: string
      phone_number:
        example: 028-87009999
        type: string
      type:
        allOf:
        - $ref: '#/definitions/_const.RecipientType'
        example: friend
      zip_code:
        example: "610072"
        type: string
    type: object
  form.BatchDeletionForm:
    properties:
      ids:
        example:
        - 5
        - 6
        - 7
        - 55
        - 124
        items:
          type: integer
        type: array
    type: object
  form.SendMailForm:
    properties:
      mail_type:
        allOf:
        - $ref: '#/definitions/_const.MailType'
        example: plain mail
      post_fee:
        example: 1.2
        type: number
      post_fee_payment_type:
        allOf:
        - $ref: '#/definitions/_const.PostFeePaymentType'
        example: stamps
      post_region:
        allOf:
        - $ref: '#/definitions/_const.PostRegion'
        example: domestic local
      recipient_address:
        example: Caotanglu 1 Hao
        type: string
      recipient_district:
        example: Sichaunsheng Chengdushi Qingyangqu
        type: string
      recipient_name:
        example: Jimmy Carter
        type: string
      recipient_phone_number:
        example: "13900000001"
        type: string
      recipient_type:
        allOf:
        - $ref: '#/definitions/_const.RecipientType'
        example: friend
      recipient_zip_code:
        example: "610061"
        type: string
      send_date:
        example: "2023-05-10T12:00:00Z"
        type: string
    type: object
  m.ContactVo:
    properties:
      address:
        type: string
      district:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      type:
        $ref: '#/definitions/_const.RecipientType'
      zip_code:
        type: string
    type: object
  m.SentMailVo:
    properties:
      id:
        type: integer
      mail_type:
        $ref: '#/definitions/_const.MailType'
      post_fee:
        type: number
      post_fee_payment_type:
        $ref: '#/definitions/_const.PostFeePaymentType'
      post_region:
        $ref: '#/definitions/_const.PostRegion'
      recipient_address:
        type: string
      recipient_district:
        type: string
      recipient_name:
        type: string
      recipient_phone_number:
        type: string
      recipient_type:
        $ref: '#/definitions/_const.RecipientType'
      recipient_zip_code:
        type: string
      send_date:
        type: string
    type: object
  resp.None:
    properties:
      _README!!!!:
        example: 如果您看到这行文字，表示这里的示例 data 字段实际是 null，这里的内容是为上层封装做注释。
        type: string
      code:
        example: 0-成功，1-失败
        type: string
      data:
        example: 实际是 null
        type: string
      msg:
        example: msg 字段显示相应错误的信息，如数据库查询失败、参数校验失败等
        type: string
      success:
        example: 成功会显示成 true，失败会被设置为 false
        type: string
    type: object
  resp.Pack-resp_None:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/resp.None'
      msg:
        example: msg
        type: string
      success:
        example: true
        type: boolean
    type: object
  resp.Pack-resp_PagingPack-m_ContactVo:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/resp.PagingPack-m_ContactVo'
      msg:
        example: msg
        type: string
      success:
        example: true
        type: boolean
    type: object
  resp.Pack-resp_PagingPack-m_SentMailVo:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/resp.PagingPack-m_SentMailVo'
      msg:
        example: msg
        type: string
      success:
        example: true
        type: boolean
    type: object
  resp.PagingPack-m_ContactVo:
    properties:
      list:
        items:
          $ref: '#/definitions/m.ContactVo'
        type: array
      page:
        example: 10
        type: integer
      page_size:
        example: 1
        type: integer
      total_pages:
        example: 3
        type: integer
      total_rows:
        example: 24
        type: integer
    type: object
  resp.PagingPack-m_SentMailVo:
    properties:
      list:
        items:
          $ref: '#/definitions/m.SentMailVo'
        type: array
      page:
        example: 10
        type: integer
      page_size:
        example: 1
        type: integer
      total_pages:
        example: 3
        type: integer
      total_rows:
        example: 24
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/v1/contacts:
    get:
      description: List contacts
      parameters:
      - description: page number, start from 1.
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Pack-resp_PagingPack-m_ContactVo'
      summary: List Contacts
      tags:
      - contact
    post:
      description: send mail
      parameters:
      - description: add contact form
        in: body
        name: HTTP_POST_body
        required: true
        schema:
          $ref: '#/definitions/form.AddContactForm'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resp.Pack-resp_None'
      summary: Add contact
      tags:
      - contact
  /api/v1/contacts/{id}:
    delete:
      description: Delete contact
      parameters:
      - description: contact ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Pack-resp_None'
      summary: Delete Contact
      tags:
      - contact
  /api/v1/contacts/any/batch-deletions:
    post:
      description: Delete contacts in batch
      parameters:
      - description: batch deletion form
        in: body
        name: HTTP_POST_body
        required: true
        schema:
          $ref: '#/definitions/form.BatchDeletionForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Pack-resp_None'
      summary: Delete Contacts In Batch
      tags:
      - contact
  /api/v1/sent-mails:
    get:
      description: List sent mails
      parameters:
      - description: page number, start from 1.
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Pack-resp_PagingPack-m_SentMailVo'
      summary: List Sent Mails
      tags:
      - mail
    post:
      description: send mail
      parameters:
      - description: send mail form
        in: body
        name: HTTP_POST_body
        required: true
        schema:
          $ref: '#/definitions/form.SendMailForm'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resp.Pack-resp_None'
      summary: Send mail
      tags:
      - mail
  /api/v1/sent-mails/{id}:
    delete:
      description: Delete sent mail
      parameters:
      - description: sent mail ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Pack-resp_None'
      summary: Delete Sent Mail
      tags:
      - mail
  /api/v1/sent-mails/any/batch-deletions:
    post:
      description: Delete sent mails in batch
      parameters:
      - description: batch deletion form
        in: body
        name: HTTP_POST_body
        required: true
        schema:
          $ref: '#/definitions/form.BatchDeletionForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Pack-resp_None'
      summary: Delete Sent Mails In Batch
      tags:
      - mail
swagger: "2.0"